+++ includes/elementor/elementor-tag-packageprice.php
@@ -1,4 +1,9 @@
 <?php
+
+use Elementor\Controls_Manager;
+use Elementor\Core\DynamicTags\Tag;
+use ElementorPro\Modules\DynamicTags\Module;
+
 if(!class_exists('\Elementor\Core\DynamicTags\Tag')){
 	die();
 }
@@ -20,7 +25,89 @@ class Elementor_Tag_PackagePrice extends \Elementor\Core\DynamicTags\Tag {
 		return [ ElementorPro\Modules\DynamicTags\Module::TEXT_CATEGORY ];
 	}
 
+
+	private function _select_accommodation(){
+
+		$accommodations = [];
+
+		foreach( get_posts([
+			'post_type' => 'accommodation',
+			'posts_per_page' => 100,
+		]) as $accommodation){
+
+			if(function_exists('get_field')){
+				$accommodation_codename = esc_html(get_field('codename', $accommodation->ID));
+			}
+
+			$accommodations[ $accommodation_codename ] = sprintf( __( 'Price for: %s', 'tmsm-woocommerce-booking-thalasso' ),
+				$accommodation->post_title );
+		}
+
+		return $accommodations;
+	}
+
+	private function _select_not_accommodation(){
+
+		$notaccommodation = [
+			'BEST' => __( 'Best Price With Accommodation', 'tmsm-woocommerce-booking-thalasso' ),
+			'TMS'  => __( 'Price Without accommodation', 'tmsm-woocommerce-booking-thalasso' ),
+		];
+
+		return $notaccommodation;
+	}
+
+	protected function _register_controls() {
+
+		$this->add_control(
+			'accommodation',
+			[
+				'label'   => __( 'Price', 'tmsm-woocommerce-booking-thalasso' ),
+				'type'    => Controls_Manager::SELECT,
+				'options' => array_merge( self::_select_accommodation(), self::_select_not_accommodation() ),
+				'default' => 'BEST',
+			]
+		);
+
+		$this->add_control(
+			'from',
+			[
+				'label'   => __( 'From Label', 'tmsm-woocommerce-booking-thalasso' ),
+				'type'    => Controls_Manager::SELECT,
+				'options' => [
+					'0' => __( 'Without From Label', 'tmsm-woocommerce-booking-thalasso' ),
+					'1' => __( 'With From Label', 'tmsm-woocommerce-booking-thalasso' ),
+				],
+				'default' => '1',
+			]
+		);
+
+		$this->add_control(
+			'instead',
+			[
+				'label'   => __( 'Instead Price', 'tmsm-woocommerce-booking-thalasso' ),
+				'type'    => Controls_Manager::SELECT,
+				'options' => [
+					'0' => __( 'Without Instead Price', 'tmsm-woocommerce-booking-thalasso' ),
+					'1' => __( 'With Instead Price', 'tmsm-woocommerce-booking-thalasso' ),
+				],
+				'default' => '0',
+			]
+		);
+
+	}
+
 	public function render() {
+
+		$from = $this->get_settings( 'from' );
+		$instead = $this->get_settings( 'instead' );
+		$accommodation_codename = $this->get_settings( 'accommodation' );
+
+		$accommodation = get_posts([
+			'post_type' => 'accommodation',
+			'meta_key' => 'codename',
+			'meta_value' => $accommodation_codename,
+		]);
+
 		$output = '';
 
 		$package = get_post();
@@ -54,7 +141,7 @@ class Elementor_Tag_PackagePrice extends \Elementor\Core\DynamicTags\Tag {
 		}
 
 		$shortcode = '[resaweb_load package_id="'.$package_idresaweb.'" lang="'.$lang.'" nights="'.$defaultnights.'"]';
-		$shortcode .= '[resaweb_price from="1" hotel_id="BEST" package_id="'.$package_idresaweb.'" lang="'.$lang.'"]';
+		$shortcode .= '[resaweb_price from="'.$from.'" instead="'.$instead.'" '.(!empty($accommodation) ? 'nights="'.$defaultnights.'"' :'' ).' hotel_id="'.$accommodation_codename.'" package_id="'.$package_idresaweb.'" lang="'.$lang.'"]';
 		$output .= do_shortcode($shortcode);
 
 		echo $output;