+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * **Tweak** - Refactoring post/body classes
+* **Tweak** - Package price with option to select accommodation, for best or without accommodation
 
 ### 1.1.8: September 21th, 2018
 * **New** - Elementor Dynamic Tag: Criteo Product ID